Arquivo atualizado com os estilos do bootstrap

+++ pdo/index.php
@@ -11,7 +11,7 @@ $sql_count = "SELECT COUNT(*) AS total FROM users ORDER BY name ASC";
 
 // SQL para selecionar os registros
 $sql = "SELECT id, name, email, gender, birthdate "
-        . " FROM users ORDER BY name ASC";
+    . " FROM users ORDER BY name ASC";
 
 // conta o total de registros
 $stmt_count = $PDO->prepare($sql_count);
@@ -24,24 +24,30 @@ $stmt->execute();
 ?>
 <!doctype html>
 <html>
-    <head>
-        <meta charset="utf-8">
-        <title>Sistema de Cadastro</title>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-    </head>
-    <body>
-        <div class="container">
+
+<head>
+    <meta charset="utf-8">
+    <title>Sistema de Cadastro</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
+</head>
+
+<body>
+    <div class="row align-items-md-stretch">
+        <div class="h-100 p-5 text-bg-dark">
             <h1>Sistema de Cadastro</h1>
-            <p><a href="form-add.php">Adicionar Usuário</a></p>
+            <h3>Manutenção de Usuário</h3>
+            <a class="btn btn-primary" href="form-add.php">Adicionar Usuário</a>
         </div>
-        <div class="container">
+    </div>
+
+    <div class="container">
         <h2>Lista de Usuários</h2>
-        <p>Total de usuários: <?php echo $total ?></p>
-        </div>
-        <div class="container">
-        <?php if ($total > 0): ?>
+        <p class="h-100 p-2 text-bg-dark rounded-3">Total de usuários: <?php echo $total ?></p>
+    </div>
+    <div class="container">
+        <?php if ($total > 0) : ?>
             <table class="table " width="50%" border="1">
-                <thead class="table-dark">
+                <thead class="table-ligth">
                     <tr>
                         <th>Nome</th>
                         <th>Email</th>
@@ -52,17 +58,16 @@ $stmt->execute();
                     </tr>
                 </thead>
                 <tbody>
-                    <?php while ($user = $stmt->fetch(PDO::FETCH_ASSOC)): ?>
+                    <?php while ($user = $stmt->fetch(PDO::FETCH_ASSOC)) : ?>
                         <tr>
-                            <td><?=$user['name'] ?></td>
-                            <td><?=$user['email'] ?></td>
-                            <td><?=($user['gender'] == 'm') ? 'Masculino' : 'Feminino' ?></td>
-                            <td><?=dateConvert($user['birthdate']) ?></td>
-                            <td><?=calculateAge($user['birthdate']) ?> anos</td>
+                            <td><?= $user['name'] ?></td>
+                            <td><?= $user['email'] ?></td>
+                            <td><?= ($user['gender'] == 'm') ? 'Masculino' : 'Feminino' ?></td>
+                            <td><?= dateConvert($user['birthdate']) ?></td>
+                            <td><?= calculateAge($user['birthdate']) ?> anos</td>
                             <td>
-                                <a href="form-edit.php?id=<?=$user['id'] ?>">Editar</a>
-                                <a href="delete.php?id=<?=$user['id'] ?>" 
-                                   onclick="return confirm('Tem certeza de que deseja remover?');">
+                                <a class="btn btn-success" href="form-edit.php?id=<?= $user['id'] ?>">Editar</a>
+                                <a class="btn btn-danger" href="delete.php?id=<?= $user['id'] ?>" onclick="return confirm('Tem certeza de que deseja remover?');">
                                     Remover
                                 </a>
                             </td>
@@ -70,10 +75,13 @@ $stmt->execute();
                     <?php endwhile; ?>
                 </tbody>
             </table>
-        <?php else: ?>
+        <?php else : ?>
             <p>Nenhum usuário registrado</p>
         <?php endif; ?>
-        </div>
-        
-    </body>
+    </div>
+    <footer class="p-3 text-bg-dark fixed-bottom text-center">
+        Desenvolvido por Maromo&#174;
+    </footer>
+</body>
+
 </html>
